+++ app/page.tsx
@@ -1,5 +1,5 @@
 import { dealHands } from "../game/dealHands";
-import { PlayField } from "./components/PlayField";
+import { PlayField } from "../components/PlayField";
 import { defaultOrder } from "../game/orders";
 
 export default function Home() {
