+++ components/PlayerHand.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { FC, useState } from "react";
-import { Card } from "../../game/types";
+import { Card } from "../game/types";
 import { PlayCard } from "./PlayCard";
 import { HandOfCards } from "./HandOfCards";
